Pull out srl v5 generator into generator_bases
